Closes #13

+++ Backend/database/migrations/migrations.sql
@@ -87,7 +87,7 @@ CREATE TABLE users(
 )ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
 CREATE TABLE have_order(
-   id_article INT,
+   id_article SMALLINT,
    id_user INT,
    id_order INT AUTO_INCREMENT,
    currentdate timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -97,7 +97,7 @@ CREATE TABLE have_order(
 )ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
 CREATE TABLE have_fav(
-   id_article INT, 
+   id_article SMALLINT, 
    id_user INT,
    PRIMARY KEY(id_article, id_user),
    FOREIGN KEY(id_article) REFERENCES article(id_article),
